requirements.txt
langchain
langchain_groq
streamlit
langchain-openai
python-dotenv